refactor(e2e): promote single-node to a sibling category
- git mv src/multi-node/single-node → src/single-node (all test files).
- git mv src/multi-node/single_node_test_context.ts → src/single-node/single_node_test_context.ts.
- Fix the two import sites in multi_node_test_context.ts:
  ./single_node_test_context.js → ../single-node/single_node_test_context.js.
- Fix imports in all moved test files: ../../fixtures/ → ../fixtures/,
  ../single_node_test_context.js → ./single_node_test_context.js.
- Update bootstrap.sh discovery glob from src/multi-node/single-node/*.test.ts
  to src/single-node/*.test.ts; re-point the 15m timeout case to single-node/proving/long_proving_time.
- Re-point the three l1_reorgs entries in .test_patterns.yml (keep them skipped; add A-1266).
- Add catch-all src/single-node/.*\.test\.ts owner entry in .test_patterns.yml.
- Add src/single-node/README.md.
- Update multi-node/README.md base-class hierarchy + subfolder map.
