+++ .test_patterns.yml
@@ -267,7 +267,7 @@ tests:
       - *charlie
       - *adam
 
-  - regex: "src/multi-node/single-node/l1_reorgs.parallel.test.ts"
+  - regex: "src/single-node/l1_reorgs.parallel.test.ts"
     owners:
       - *phil
 
@@ -277,15 +277,15 @@ tests:
   # inconsistency disturbs the tip read during the prune. Under CI's L1 cadence the asserted value is
   # never observed within the wait window, so original + retry both time out and the flake retry hard-fails
   # (these pass locally at the faster L1=4 cadence). Pre-existing timing fragility — the file was previously
-  # skip:true (added in 60b274c, the skip lifted in #23642) — not introduced by the multi-node move, whose
-  # FAST_REORG_TIMING config resolves byte-identically to the prior values. Skipped per-`it` so the other
-  # five cases in the file stay covered. Owner: phil.
-  - regex: "src/multi-node/single-node/l1_reorgs.parallel.test.ts.*prunes L2 blocks if a proof is removed due to an L1 reorg"
+  # skip:true (added in 60b274c, the skip lifted in #23642; tracked in A-1266) — not introduced by the
+  # single-node promotion, whose FAST_REORG_TIMING config resolves byte-identically to the prior values.
+  # Skipped per-`it` so the other five cases in the file stay covered. Owner: phil.
+  - regex: "src/single-node/l1_reorgs.parallel.test.ts.*prunes L2 blocks if a proof is removed due to an L1 reorg"
     skip: true
     owners:
       - *phil
 
-  - regex: "src/multi-node/single-node/l1_reorgs.parallel.test.ts.*restores L2 blocks if a proof is added due to an L1 reorg"
+  - regex: "src/single-node/l1_reorgs.parallel.test.ts.*restores L2 blocks if a proof is added due to an L1 reorg"
     skip: true
     owners:
       - *phil
@@ -309,6 +309,11 @@ tests:
     flake_group_id: e2e-p2p-epoch-flakes
     owners:
       - *phil
+
+  - regex: "src/single-node/.*\\.test\\.ts"
+    flake_group_id: e2e-p2p-epoch-flakes
+    owners:
+      - *palla
 
   - regex: "src/multi-node/.*\\.test\\.ts"
     flake_group_id: e2e-p2p-epoch-flakes
